requirements.txt
torch==2.7.1
torchvision==0.22.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
python-multipart==0.0.9
scikit-learn==1.5.1
numpy==1.26.4
customtkinter==5.2.2